requirements.txt
Flask
Flask-SQLAlchemy
requests
cassandra-driver
